Merge [22725] from trunk.
Fix an issue with the incorrect status being displayed in the mythfrontend
status box for a tuner that has an error. The status was previously
indicating that the tuner was asleep when in fact there was an error
condition reported by the backend.

Closes #7609.  /me doesn't like reminder tickets for things on his TODO list,
but he didn't want to go to bed grumpy so he merged this over now anyway.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-22-fixes@22860 7dbf422c-18fa-0310-86e9-fd20926502f2