Initial commit for tmdb.py, a new TMDB Grabber script for TheMovieDB.
How does tmdb.py differ from tmdb.pl?  I'm glad you asked!

tmdb.py uses the new TMDB v2.1 API, which is being expanded with all sorts of new information.  When MPAA ratings/Parental rating is added to this new API, we will be able to add them simply and quickly, along with tons of other new info the API will be exposing.

tmdb.py takes IMDB *or* TMDB numbers as an argument, and spits out the TMDB number in the results, which means running a grab with the new grabber will seamlessly move an existing record over to TMDB numbers.

tmdb.py's search takes into account *all* films at TMDB, not just those with an IMDB number set.

tmdb.py is an alpha script.  I would prefer it if people didn't use it with .22 as a) it's unsupported and b) .22 doesn't include some key bits of grabber logic to manage the IMDB->TMDB transition, and c) .23 is out in March, take a deep breath.  It will work, I would just prefer that I (and Doug, and others on the lists and IRC channels) not have to provide support for an in progress script.

Like the famous Jamu and tvdb.py scripts, tmdb.py is written by Doug Vaughan, whose dedication to supporting his scripts and their users is always really, really appreciated.

Closes #6467.


git-svn-id: http://svn.mythtv.org/svn/trunk@22865 7dbf422c-18fa-0310-86e9-fd20926502f2