Implement more of the hashing logic for MythVideo.  Note that this is presently limited to locally stored copies, but one can now freely rename, move, spindle, or mutilate the filenames without losing metadata.  As long as the files are still the same *data*, they'll be reconnected to the metadata on a rescan.
In testing, hashing and adding 1100 files took 54 seconds.  Files are not hashed unless they are renamed or being added.  Existing files are not rehashed on a scan.  That's the long way of saying it takes a little longer for the initial add, but that there's no performance hit from having a lot of records in subsequent searches.

Looking at how best to work this with Storage Group files, hope to have something for that in a few days time.


git-svn-id: http://svn.mythtv.org/svn/trunk@22870 7dbf422c-18fa-0310-86e9-fd20926502f2