Update MythVideo to use the generic file hashing method for all content, whether stored locally or in Storage Groups.  This means that any content in MythVideo can be freely moved between storage groups, between local and SG storage, or any permutation thereof, and a simple rescan will reconnect the file locations with the metadata.
Note that hashing is *only* done on files that are a) eligible for delete or b) eligible for insertion.  Thus, the speed of a scan of the existing content is unaffected.  There is a very small slowdown to hash the file when the file is being added, but tests here indicate that around 15 hashes/sec on a Q6600 are the normal speed.  

This commit includes a DB update to hash all existing material.  Depending on the speed of your processor and the size of your collection, the update could take a few moments.  The hashing progress is printed in a general verbose message.


git-svn-id: http://svn.mythtv.org/svn/trunk@22887 7dbf422c-18fa-0310-86e9-fd20926502f2