Refs #7696. Reconnect frontend to backend after disconnects.
This normally happens immediately after you try to do something in the frontend after a disconnect, but there are two problems. First, this can take a couple seconds, this gives the frontend the appearance of slugishness. Second, since we didn't know of the disconnect we didn't know to reload our entries in the Watch Recordings screen.

We may want to do some tweaking of the timeouts, but I want to address #7696 first; trac wasn't behaving so I couldn't look at the ticket tonight...

git-svn-id: http://svn.mythtv.org/svn/trunk@22958 7dbf422c-18fa-0310-86e9-fd20926502f2