Refs #7691. Don't update pathname when we get a ProgramInfo update event, these could may have been localized.
Ideally we should keep a pristine pathname in each ProgramInfo and update a non-serialized pathname with any locallized version. Until we reach that level of cleanliness in ProgramInfo this hack will avoid problems locating the file when we expect it to be localized to the reciever of the update event.

git-svn-id: http://svn.mythtv.org/svn/trunk@22968 7dbf422c-18fa-0310-86e9-fd20926502f2