After [22970], remove the commented <LocalHostName> element since the Perl bindings do not use a real XML parser to read config.xml, so they were picking up the commented value.
I hope to revert this change once I get approval to modify the bindings to use XML::Simple or XML::Parser for reading config.xml.  While I think the user of a real parser was avoided so as not to add a dependency, I can't imagine a system having Perl, but lacking XML::Simple and/or XML::Parser (or at the very least, expat), these days.  I believe the added robustness for parsing config.xml is worth the requirement (since config.xml is often edited by users, who may not realize that commented elements will still be used by the Perl bindings).


git-svn-id: http://svn.mythtv.org/svn/trunk@22972 7dbf422c-18fa-0310-86e9-fd20926502f2