Fix for recstatus update with UPDATE_PROG_INFO conversion.
The problem was that the recstatus is pulled specially from the scheduler when upding a whole list of recordings at once. This adds a method to the scheduler to allow pulling a single recstatus. This also introduces two new MythProto commands, MASTER_UPDATE_PROG_INFO and UPDATE_PROG_FILESIZE. The first is so we can send the update to the master first so it can pull the updated recstatus before sending out UPDATE_PROG_INFO to the frontends. The second makes updating just the filesize more efficient, we update that many more times than we update other components of the ProgramInfo.

There is another problem with playbackbox not updating the color of text in response to a ProgramInfo update. I've worked around it in this commit by resetting the UI text with the updated status, but stuart m will be committing a more efficient solution a little bit later (this workaround is still about 500 times faster than the old code...)



git-svn-id: http://svn.mythtv.org/svn/trunk@22976 7dbf422c-18fa-0310-86e9-fd20926502f2