Refs #7603. This moves the QUERY_RECORDINGS mythproto calls out of the UI thread.
The idea is to move as many mythproto calls as possible out of the UI thread so that when a reconnect to the master backend is necessary we can show a dialog to that effect during the reconnect attempt even if using MythUI.

QUERY_RECORDINGS is really the major query that this screen does. But there are also some done when parsing the list returned by this call, which I'll tackle next. And various mythproto calls linked to various actions that don't require the return value, which I'll tackle as one task with a QRunnable dispatcher.

git-svn-id: http://svn.mythtv.org/svn/trunk@23009 7dbf422c-18fa-0310-86e9-fd20926502f2