Speeds up the handling of no-file and zero-byte recordings on the backend
While testing some code I generated a lot of no-file and zero-byte recordings and discovered that we're spending an inordinate amount of time checking those do to a subtle interaction with the storage group code.

The ProgramInfo::GetPlaybackURL() was returning "/GetPlaybackURL/UNABLE/TO/FIND/LOCAL/FILE/ON/..." for files it couldn't find, but we have a few places in the code that assume anything starting in "/" is a valid file. This strips the leading "/" from the message, and changes a couple other checks.



git-svn-id: http://svn.mythtv.org/svn/trunk@23042 7dbf422c-18fa-0310-86e9-fd20926502f2