Look up programs that can not yet be found in the recorded table in the program table, failing that make sure the format of the %STARTTIME% is the same.
We were returning the starttime in ISO format until the recording was inserted into the DB, at which point we switched to mythtv format.

The starttime isn't really accurate until the recording starts, so listening to the pending recording event still doesn't give you the exact recstartts, but having it in the same format as the sent by the started recording script makes it easier to use.

Looking up the program in program for the pending recording events just allows you to log the title and other niceties for humans.



git-svn-id: http://svn.mythtv.org/svn/trunk@23043 7dbf422c-18fa-0310-86e9-fd20926502f2