Refs #7603. Moves the available status checking to the helper thread and simplifies some of the popup code.
Note: This is the last of the MythProto stuff in playbackbox.cpp aside from findArtworkFile(), themes not using findArtworkFile() should now feel a bit snappier. There are still the occational MythProto accesses in the UI thread in other cpp's used by the PBB, and there are still a number of Get*Setting() calls which slow things down when the item is not in the cache and we need to reconnect the DB socket.

git-svn-id: http://svn.mythtv.org/svn/trunk@23047 7dbf422c-18fa-0310-86e9-fd20926502f2