Refs #7603. This makes the reselection code when reloading a the full list a bit smarter and cleans up verbose error and warning macro usage.
The reselection code now attempts not just the previously selected item, which may have been deleted or moved to another recording group which is not currently visible, it also tries the following items in the same group, then the previous items in the same group, and if the group is gone it tries the following groups.

This should avoid a lot of the jumps to the top of the 'All Recordings' group that happen when the currently selected item is deleted, either locally or on another frontend.

git-svn-id: http://svn.mythtv.org/svn/trunk@23048 7dbf422c-18fa-0310-86e9-fd20926502f2