Refs #7603. Fixes segfault in mythcontext popup handling.
This drops the old UI popups and uses only MythUI pop-ups. The MythUI pop-ups do not try to launch from the calling thread anymore but instead an event is dispatched to an object whose event handler is in the MythUI thread. This simplifies the code and avoids all the problems, including segfaults, which occur when you try to use MythUI outside the MythUI thread.

Note: When MythProto is used within the MythUI thread, the popup will not be shown when it should be. This is pre-existing behaviour and is inherent in the design of MythUI. But a previous commit causes such usage of MythProto to print a warning message:
SendReceiveStringList(...) called from UI thread
These should be fixed, as the have potential to make the frontend appear to be hung for minutes at a time.

git-svn-id: http://svn.mythtv.org/svn/trunk@23051 7dbf422c-18fa-0310-86e9-fd20926502f2