Remove the StickyKeys setting as discussed in the Settings Rework thread ( http://www.gossamer-threads.com/lists/mythtv/users/405626#405626 ).
Enabling the setting actually removes functionality from MythTV.  All it does is make the SEEKFFWD work like FFWDSTICKY and SEEKRWND like RWNDSTICKY, so once enabled, the user no longer has an action to use for SEEKFFWD or SEEKRWND.  Therefore, users wanting to always use the sticky forward/rewind should instead bind the desired keys to the FFWDSTICKY/RWNDSTICKY actions or change remote definitions to use the appropriate keys for FFWDSTICKY (> or .) and RWNDSTICKY (< or ,).

As a part of the change, a DB update is included which will unbind Left/Right from SEEKRWND/SEEKFFWD and bind them to RWNDSTICKY/FFWDSTICKY if the StickyKeys setting is enabled for the host and all 4 bindings use the default keylists.  Therefore, the change should be transparent to nearly all users.

As far as I can tell from our sparse pre-SVN records, the setting was added around Oct 2002 ( http://www.gossamer-threads.com/lists/mythtv/dev/464#464 ).  When the FFWDSTICKY and RWNDSTICKY actions were added in [2568] (Nov 2003), they obsoleted the old StickyKeys setting.

This change also updates the DB Schema Version for the Python bindings and adds a comment to dbcheck.cpp to remind us (me?) to update the bindings when currentDatabaseVersion changes (as I forgot to do so with DB update 1249).


git-svn-id: http://svn.mythtv.org/svn/trunk@23112 7dbf422c-18fa-0310-86e9-fd20926502f2