Make scaled preview image filesystem permissions match those of the "default-sized" preview images.
This change basically applies the same preview-image chmod 666 hack added in [9566] and [9571] to the preview images created by MythXML.  Rather than replicate the hack, yet again, it moves the chmod to makeFileAccessible() in libs/libmyth/util.cpp and updates all the previous code to call that function.

The patch also removes some includes that no longer seem to be necessary after removing the calls to chmod().  Please send an e-mail to the list or mention on IRC if your compiler disagrees with me/my compiler.

Closes #4860 with patch from sphery.


git-svn-id: http://svn.mythtv.org/svn/trunk@23113 7dbf422c-18fa-0310-86e9-fd20926502f2