Removed the ComplexPriority scheduler setting and added a feature to warn when a less preferred showing of a program might be recorded.
Removing the the ComplexPriority setting was suggested in the settings
thread over a month ago.  The setting was originally added to address
some deficiencies in the "Reschedule Higher Priorities" logic in the
scheduler.  Those deficiencies were mostly obviated by a later change
making this setting unnecessary.

The new feature attempts to detect when a conflict causes one showing
of a program to be recorded instead of another showing with a higher
priority.  For example, if the user has priorities setup to prefer HD
recordings and tuners, this could warn when a program gets bumped to
an SD tuner.  Note that this feature currently only checks showings at
the same time.  It does not detect if there are higher priority
showings at other times.


git-svn-id: http://svn.mythtv.org/svn/trunk@23118 7dbf422c-18fa-0310-86e9-fd20926502f2