Add MythNetvision as a BETA plugin.  Note the Beta status, where it will remain for .23 and possibly for a little bit afterwards depending on progress.
That said, MythNetvision is the beginnings of an (official!) Internet Video plugin for MythTV.  It features a fully themeable MythUI component, and uses user contributed scripts to parse and return information for various video sites.  When possible, MythNetvision will offer options to download the video to your hard drive.  When not possible, you can view the content in a browser (MythBrowser preferred).

MNV contains two components-- an Internet video search, and a tree/site view browser.  The site browser is analogous to MythVideo in that it has three themed views.  MNV also supports RSS feeds, since the core of the grabber scripts is the RSS media format.

MNV is very, very, very far from perfect and complete.  It has the following major known issues:

1) It is possible to crash the plugin if you browse too quickly between directories at the exact moment a signal is send back from the thumbnail downloader (critical).
2) It is currently necessary to invoke a refresh of the RSS and Sites manually from the menu. (User Experience/High)
3) The downloader functionality is just a stub.  Right now, a downloadable file just gets dumped into the working directory for the frontend, and played when the download is complete.  The optimal solution to this is to trigger a download by the backend where it is inserted into recordings by an IPTV recorder, which would also allow in-progress downloads to play back. (Medium/possibly not doable in time for .23)
4) The RSS Feed updates require two refreshes of the tree to show up, either by leaving the plugin or running the update twice.  It has to do with the update not being done by the time the tree gets reloaded (User Experience/Low).
5) The included complement of directory icons are way too large sizewise and need to be replaced by pretties equivalents.  This would be a very welcome improvement from those of you who are artistically minded!

With all of the above in mind, please be patient with me/us as this progresses.  I am eager to have you all involved in improving and developing this plugin, but it is definitely taxing my meager skills.  By throwing it open for us all to work on (that includes you, community/users!), it's my hope that we can polish it up by .23 and make it a great platform to build on in the future.  Patches to address *any* issue, especially those found above would be very, very welcome.

The MythNetvision grabbers included are written by Doug Vaughan, who has also written a framework upon which new grabbers can be easily added (see mythnetvision/scripts/nv_python/libs).  The current included scripts are:

DailyMotion
TheMovieDB Trailers
YouTube
Joost
Vimeo
MTV
Blip.tv

I have also received several other grabbers over the past few days via e-mail.  Once we clear up some of the issues with those and verify ToS, it's my hope that we'll be adding several more grabbers to this list.

Enjoy, and I really look forward to improving the plugin further over the next few months.


git-svn-id: http://svn.mythtv.org/svn/trunk@23123 7dbf422c-18fa-0310-86e9-fd20926502f2