Remove the "Show Group Summary" setting widget.  The setting has been disabled s ince [19603] (the playbackbox port to MythUI).
In the settings rework thread, removal of the Show Group Summary setting was proposed since it's no longer used.  We got 2 comments expressing a desire to keep the group summary itself, but no one made any case for allowing the group summary to be disabled.  Even though the setting itself is not actually used (and hasn't been for a year), the current behavior is the behavior both responders wanted:  MythTV always shows information about the selected group when in the groups list.  Therefore, this change removes the no-longer-used setting widget.

Thanks to Tim Mackinnon for noticing the old widget and reminding us to remove it.  Refs #7905.


git-svn-id: http://svn.mythtv.org/svn/trunk@23128 7dbf422c-18fa-0310-86e9-fd20926502f2