Improves reselection code in playbackbox. Try to not just keep the same item selected on reloads, but also the same scroll position.
We don't try as hard to keep the scroll position near where it was as we do with the selection. The reasoning being that if the list has changed drastically we should really by scrolling to a new position. We attempt to locate the former top item and the item above and below it, failing that we allow the list to scroll without restraint.

git-svn-id: http://svn.mythtv.org/svn/trunk@23139 7dbf422c-18fa-0310-86e9-fd20926502f2