Makes two optimizations to SetPositionMap[Delta]
The first is to only call prepare once for the many inserts (and only bind the non-changing things once).
The second is to use INSERT DELAYED instead of just a regular INSERT.

The first is always safe and saves lots of time with remote DB's, the second delays inserts if a read on the table is in progress, I did not find that this apprecably delayed the inserts in my tests, but if someone finds that it does with their preferred DB engine we can revert this part.

For the second, INSERT DELAYED is also not SQL-92 compliant, but it can be safely removed if we move to a different database sql frontend.



git-svn-id: http://svn.mythtv.org/svn/trunk@23150 7dbf422c-18fa-0310-86e9-fd20926502f2