Speeds up NVP::RebuildSeekTable while also saving the position map during the rebuild.
Instead of saving the position map in one fell swoop at the end of the rebuild, this instead incrementally updates it. If we just did this in the rebuild thread we'd be blocking while the mysql libs wait for Nagle timeouts to flush data (with remote DB's), so instead we use the threadpool to do the DB updates. We also use ProgramInfo::SetPositionMapDelta instead of SetPositionMap, since we already clear the pre-existing position map at the start of the rebuild.

git-svn-id: http://svn.mythtv.org/svn/trunk@23151 7dbf422c-18fa-0310-86e9-fd20926502f2