Fixes #7436.  Remove a couple of duplicate indices.
The recgroup index on the recgrouppassword table was added in [3112] when a named index identical to the primary key index was created during table creation, likely just an oversight.  This duplicate index should exist in all valid mythconverg schemas, and would only be missing if users edited their schemas (which they should never do).

The title_2 index on the videometadata table was added in [2635] as a quick solution to fixing a broken DB upgrade path.  This duplicate index is likely to exist on most mythconverg schemas, but it is possible that it won't exist in some valid mythconverg schemas (that were created long ago) due to the broken DB upgrade path fixed by [2635].

Though neither index caused any harm, they're being removed only for cleanliness of the schema design.  There are no functional changes due to this update.  In all, the end result is that the removal of the recgroup index may save up to 1 kiB and the removal of the title_2 index may save up to 10's of kiB of storage space on the hard drive.

Checking to see if the index exists before removing it simply prevents logging errors that aren't errors and, even if there is an error removing the index, the error is ignored (as the extra indices don't cause any harm) and the schema update is considered complete and successful.


git-svn-id: http://svn.mythtv.org/svn/trunk@23202 7dbf422c-18fa-0310-86e9-fd20926502f2