Use QEvent::registerEventType() to generate unique event id's for MythTV events.
This allows one to add new events without having to search the source to try to find a user event id not currently in use. The registerEventType() method was introduced in Qt4.4 to avoid having to do that...

I have not tested every single event, but tried to get good coverage and used the same transformation throughout the code.

git-svn-id: http://svn.mythtv.org/svn/trunk@23211 7dbf422c-18fa-0310-86e9-fd20926502f2