Fix sizing and positioning of DVD and DVB subtitles and consolidate with DVD menu button sizing code.
This appears to be pretty complete and seems to handle just about every
combination of OSD type, source material, screen adjustments, overscan
and aspect/zoom override that I can test with.

There is a cropping problem with certain, more extreme zooms but this is
a problem in the OSD code that I'll address separately.

Refs r23209, r23189.


git-svn-id: http://svn.mythtv.org/svn/trunk@23220 7dbf422c-18fa-0310-86e9-fd20926502f2