Fix for plugin name command-line argument parsing with a better fix than the fix in [23208].  This fix seems to be more in line with the original design of the command-line parser and doesn't cause a failure when the final argument in a command line is not a plugin name but is a valid argument that looks like a plugin name (doesn't start with -, such as the second argument of a 2-part argument, like -v or -O).
Note that if you pass an invalid plugin name for what is meant to be the pluginname (like mythvidoe), the frontend will still be executed, but without any menus or UI showing, and will require interrupting the program.  This also means that the fallback to adding "myth" before the argument value won't work (so, while it should work, you can't use video as the plugin name).  This will be fixed separately.


git-svn-id: http://svn.mythtv.org/svn/trunk@23225 7dbf422c-18fa-0310-86e9-fd20926502f2