Various MythWelcome updates. Refs #7725.
Move the default 4:3 mythwelcome theme file into the default theme with
all the other theme files and remove the mw_* button images. 

Update the default and default-wide themes to us a clock widget for both the 
time and date and remove the code which was updating the date and time.

Use the basewidebutton for the button rather than a custom one only used
by MythWelcome.



git-svn-id: http://svn.mythtv.org/svn/trunk@23248 7dbf422c-18fa-0310-86e9-fd20926502f2