Improved AVFormatDecoder seeking with patches from Taylor Ralph.
I've been testing for a few days and the improvement in navigating files
without a markup (i.e. MythVideo) is substantial.

I'd delayed due to an issue seen with certain test matroska files but
I've now realised that this is a separate, existing seek problem with
these files that manifests itself slightly differently following these
changes.

Closes #7521, Closes #7935.


git-svn-id: http://svn.mythtv.org/svn/trunk@23288 7dbf422c-18fa-0310-86e9-fd20926502f2