Modify mythtranscode such that it doesn't attempt to transcode files when --buildindex is used.  It also prevents mythtranscode from generating a seek table for Matroska file (since the presence of the seek table broke playback of some files until the seek table was cleared, and libav* seeking in MKV files is rather good in the Internal player).
Since mythtranscode will build a seek table after transcoding recording files, the use of --buildindex implies that no transcoding is being performed.  Users who have used --buildindex along with a transcode command for MythVideo video files may need to explicitly run a separate mythtranscode --buildindex command after the transcode (if they want a video seek table).

This patch changes the application flow sufficiently that we no longer enter the code block that was causing the 0-byte .tmp file when running a --buildindex, as reported on #5110.  Therefore, this also fixes that issue.

Thanks to Simon Walls for the patch on #5110 (which was correct, but is OBE by this change) and to Taylor Ralph for this patch.

Fixes #7986.  Fixes #5110.


git-svn-id: http://svn.mythtv.org/svn/trunk@23363 7dbf422c-18fa-0310-86e9-fd20926502f2