Cleanup MythCenter[-wide] some more.
The recgroups selector didn't really fit in, instead of presenting it as a list, this makes just the current recording group visible, without that it's difficult to tell if you are in 'Default' or 'All Programs', and sometimes even other large groups.

There were also some font offset issues with MythCenter-wide I noticed at different resolutions. Ideally we should fix the font metrics compution in MythUI so that we don't need offsets, but that shouldn't happen just before a release as it will break all existing themes. For now, I've just picked some compromise offsets so we stay within a couple pixels of the desired value on one side or the other.

git-svn-id: http://svn.mythtv.org/svn/trunk@23390 7dbf422c-18fa-0310-86e9-fd20926502f2