Don't try and switch renderer if the video resolution is unchanged.
VideoOutput::IsPreferredRender() was originally designed to check that
the user's chosen renderer was being used for a given video resolution.
I failed to consider that the current renderer may not be the original
choice for that resolution (e.g. broken profile, failed VideoOutput
initialisation etc) and hence all types of input change (e.g. aspect
ratio) would trigger a re-initialisation. If the current renderer is
displaying the current video resolution, there is probably a good reason
for it!

Implicated in #7967 and #7779 and no doubt a few other tickets.


git-svn-id: http://svn.mythtv.org/svn/trunk@23398 7dbf422c-18fa-0310-86e9-fd20926502f2