Delay CEA-708 caption window deletes until there is something to replace em.
FOX sends a DeleteWindow command almost immediately after a caption is displayed so that you need to really concentrate to catch every caption. This commit delays the deletion of an old caption window until the next one is displayed.

This can leave a caption on screen for some time if there is a long silence in the program, but I think this is preferable to unreadable captions.

git-svn-id: http://svn.mythtv.org/svn/trunk@23411 7dbf422c-18fa-0310-86e9-fd20926502f2