Fix a regression in the JobQueue from [21369].  During the QMap key conversion from chanid+starttime to jobID, the JobQueue lost the ability to check for another job running for the same recording and multiple jobs could be running at the same time for a given recording.  This patch fixes this regression so that multiple jobs for a given recording are run serially instead of in parallel.
Also add a locking in some places around reads of the runningJobs QMap.
These haven't been there since the beginning of the JobQueue, but we
need to protect our reads to make sure we're not reading while the
QMap is being updated in another thread.

References #7731 since the regression above was mentioned in this ticket.



git-svn-id: http://svn.mythtv.org/svn/trunk@23415 7dbf422c-18fa-0310-86e9-fd20926502f2