Platform dependant VideoDisplayProfile options.
We now build the list of supported decoders, renderers and associated
options from the decoder and video output classes at run time. Any
unsupported options are ignored when building the profile list from the
database entries.

The net result is that the user is only presented with options that have
been compiled in and the playback code will not try to initialise
renderers that are unavailable. As before, final selection and
validation of decoder/renderer support is completed as needed (e.g. XvMC
support for MPEG2..)

Closes #7967, Refs #7972.


git-svn-id: http://svn.mythtv.org/svn/trunk@23420 7dbf422c-18fa-0310-86e9-fd20926502f2