Remove the old, legacy themes.
This changes leaves the trunk/themes directory in place as a placeholder for forthcoming themes (such as those from the theming competition), though there are no themes in it.  Currently the configure script will print a message and exit (with error) and, if a user ignores the message and gets to make install, it will do nothing.

Though trunk/oldthemes/Retro has an outstanding ticket with updates for mythui (#7533), the ticket changes Retro to Retro-wide and includes a complete tarball of the theme, so there's no need to keep the current shadow-of-the-old-Retro-theme.


git-svn-id: http://svn.mythtv.org/svn/trunk@23442 7dbf422c-18fa-0310-86e9-fd20926502f2