Refs #7913. Fixes bug in [23402] identified by Jim Stichnoth.
TTFFont contains not only the font. It also contains the color, shadow and outline info. After [23402]
any two TTFFont's that shared the same font would always have the same color, shadow and outline state.
This fixes that by seperating out the font info from the other info. The font info itself is refernce
counted and cleaned up once the last TTFFont referring to it is deleted.



git-svn-id: http://svn.mythtv.org/svn/trunk@23460 7dbf422c-18fa-0310-86e9-fd20926502f2