Improved VDPAU thread safety.
Firstly, internalise all access to the display device in MythRenderVDPAU
to ensure calls are serialised. Secondly, extend and tidy up locking
within VideoOutputVDPAU to ensure access to the MythRenderVDPAU object
is always safe when the object is being re-created. Generally this means
applying the lock in all public methods (hence assuming that private
methods are safe).

The only exception to this is VideoOutputVDPAU::DrawSlice. Locking here
would synchronise the decode and render threads (not a good thing) but
actually the main issue being addressed here is when the decoder thread
initiates recreation of the object when the input changes - hence calls
to InputChanged and DrawSlice should not be concurrent.

Closes #7779.


git-svn-id: http://svn.mythtv.org/svn/trunk@23467 7dbf422c-18fa-0310-86e9-fd20926502f2