Modify mythlink.pl to allow creating a symlink for a specific recording file using the --filename or --chanid and --starttime arguments.  This allows the script to be used as a proper user job or--even better--to be used in conjunction with the "Recording started" event in the new event-driven notification system so links are created as soon as the recordings start.
Note that when specifying --filename or --chanid and --starttime, deletion of old symlinks (other than the one for the specific file) is suppressed.  Therefore, anyone using this script in a user job or with the event system should still run the script occasionally without the --filename or --chanid and --starttime arguments to allow removal of links that referred to recordings since deleted.

The --filename argument can use either the absolute path of the file or the recording file's basename.  The --starttime argument can use either the MythTV time format (yyyy-MM-ddThh:mm:ss), the ISO time format (yyyy-MM-dd hh:mm:ss), or the "job queue" time format (yyyyMMddhhmmss).  The --chanid uses the channel ID format.

Thanks to Johan Godfried for the initial patch.  Fixes #5379 with a slightly modified version of the patch on the ticket.


git-svn-id: http://svn.mythtv.org/svn/trunk@23479 7dbf422c-18fa-0310-86e9-fd20926502f2